# docs/docs/swagger.yaml
basePath: /
definitions:
...
  internal_transport_http.AddUserDomainRequest:
    properties:
      domain:
        type: string
      mode:
        enum:
        - shared
        - exclusive
        - catch_all
        - whitelist
        type: string
    required:
    - domain
    - mode
    type: object
...
  internal_transport_http.Response:
    properties:
      code:
        description: 业务状态码
        type: integer
      data:
        description: 数据载荷
      msg:
        description: 中文提示信息
        type: string
    type: object
...
  internal_transport_http.UpdateDomainModeRequest:
    properties:
      mode:
        enum:
        - shared
        - exclusive
        - catch_all
        - whitelist
        type: string
    required:
    - mode
    type: object
...
  tempmail_backend_internal_domain.DomainMode:
    enum:
    - shared
    - exclusive
    - catch_all
    type: string
    x-enum-varnames:
    - DomainModeShared
    - DomainModeExclusive
    - DomainModeCatchAll
  tempmail_backend_internal_domain.DomainStatus:
    enum:
    - pending
    - verified
    - failed
    - expired
    type: string
    x-enum-varnames:
    - DomainStatusPending
    - DomainStatusVerified
    - DomainStatusFailed
    - DomainStatusExpired
...
  tempmail_backend_internal_domain.UserDomain:
    properties:
      createdAt:
        description: 创建时间
        type: string
      domain:
        description: 域名（例如：example.com）
        type: string
      expiresAt:
        description: 过期时间（独享模式付费到期）
        type: string
      id:
        type: string
      isActive:
        description: 是否激活
        type: boolean
      lastCheckAt:
        description: 最后检查时间
        type: string
      mailboxCount:
        description: 该域名下创建的邮箱数
        type: integer
      mode:
        allOf:
        - $ref: '#/definitions/tempmail_backend_internal_domain.DomainMode'
        description: 共享/独享模式
      monthlyFee:
        description: 月费（独享模式）
        type: number
      mxRecords:
        description: MX 记录配置
        items:
          type: string
        type: array
      notes:
        description: 备注
        type: string
      status:
        allOf:
        - $ref: '#/definitions/tempmail_backend_internal_domain.DomainStatus'
        description: 验证状态
      userId:
        description: 域名所有者
        type: string
      verifiedAt:
        description: 验证时间
        type: string
      verifyMethod:
        description: 验证方式（dns_txt/file）
        type: string
      verifyToken:
        description: DNS TXT 验证令牌
        type: string
    type: object
...
info:
  contact:
    email: support@example.com
    name: API Support
  description: TempMail 后端 API 文档
  title: TempMail Backend API
  version: 0.9.0
paths:
...
  /v1/user/domains:
    get:
      description: 获取当前用户添加的所有域名
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tempmail_backend_internal_domain.UserDomain'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
      summary: 获取用户的域名列表
      tags:
      - User Domains
    post:
      consumes:
      - application/json
      description: 用户添加自己的域名，可选择共享或独享模式
      parameters:
      - description: 域名信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http.AddUserDomainRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/tempmail_backend_internal_domain.UserDomain'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
      summary: 添加用户自定义域名
      tags:
      - User Domains
  /v1/user/domains/{id}:
    delete:
      description: 删除用户自定义域名
      parameters:
      - description: 域名ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
      summary: 删除域名
      tags:
      - User Domains
    get:
      description: 获取指定域名的详细信息
      parameters:
      - description: 域名ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tempmail_backend_internal_domain.UserDomain'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
      summary: 获取域名详情
      tags:
      - User Domains
    patch:
      consumes:
      - application/json
      description: 更新域名的共享/独享模式
      parameters:
      - description: 域名ID
        in: path
        name: id
        required: true
        type: string
      - description: 模式信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http.UpdateDomainModeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tempmail_backend_internal_domain.UserDomain'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
      summary: 更新域名模式
      tags:
      - User Domains
  /v1/user/domains/{id}/instructions:
    get:
      description: 获取域名的 DNS 配置说明（MX 记录、TXT 记录）
      parameters:
      - description: 域名ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
      summary: 获取域名配置说明
      tags:
      - User Domains
  /v1/user/domains/{id}/verify:
    post:
      description: 通过 DNS TXT 记录验证域名所有权
      parameters:
      - description: 域名ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tempmail_backend_internal_domain.UserDomain'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
      summary: 验证域名所有权
      tags:
      - User Domains
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: 兼容层 API Key
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 使用格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
  MailboxToken:
    description: 邮箱访问令牌
    in: header
    name: X-Mailbox-Token
    type: apiKey
swagger: "2.0"